dataflow tests segregation for gosrc2cpg
